feat：知识库检索完善检索逻辑
- 新增Milvus 语义向量检索
- 新增ES BM25 关键词检索
- 新增融合两路结果

+++ src/main/java/com/zxl/agi/model/MemoryItem.java
@@ -14,12 +14,12 @@ public class MemoryItem {
     private int id;
     private String content;
     private double importance;
-    private List<Double> embedding;
+    private List<Float> embedding;
     private double score;
     private LocalDateTime createdAt;
     private LocalDateTime lastAccessed;
 
-    public MemoryItem(int id, String content, double importance, List<Double> embedding) {
+    public MemoryItem(int id, String content, double importance, List<Float> embedding) {
         this.id = id; this.content = content; this.importance = importance; this.embedding = embedding;
         this.createdAt = LocalDateTime.now(); this.lastAccessed = LocalDateTime.now();
     }
